requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-classy-tags==1.0.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-modeltranslation==0.15
django-redis-cache==2.1.1
django-robots==4.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
Markdown==2.6.11
mock==4.0.2
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
